Move UI type definitions to shared folder
