+++ src/database/types.ts
@@ -1,0 +1,6 @@
+export { Edition } from './models/edition.model';
+export { GameType } from './models/gameType.model';
+export type { IPlayerStats2 } from './models/playerStats2.model';
+export type { ISeasonStats } from './models/seasonStats.model';
+export type { IPBCStats } from './models/pbcStats.model';
+export type { IMatch } from './models/match.model';
