+++ src/types/ui.types.ts
@@ -1,0 +1,16 @@
+import {
+  ActionRowBuilder,
+  ButtonBuilder,
+  StringSelectMenuBuilder,
+  EmbedBuilder,
+} from 'discord.js';
+
+export type ButtonRow = ActionRowBuilder<ButtonBuilder>;
+export type SelectRow = ActionRowBuilder<StringSelectMenuBuilder>;
+
+export interface PaginationManagerOptions {
+  prefix: string;
+  embeds: EmbedBuilder[];
+  ephemeral?: boolean;
+  timeoutMs?: number;
+}
