+++ src/ui/index.ts
@@ -1,2 +1,3 @@
 export * from './buttons';
 export * from './pagination';
+export * from '../types/ui.types';
